Imagemagic convert does not work in docker, hardcode bmp splashfile.
Could not get imagemagic convert working in reasonable time, since
this is not going to change very often, one can generate a new image
when it is needed.

For this reason the function with the commands to execute is left in the
script. It is better then reading the ImageMagic documentation (trust me).

Also left the Old splash svg in, because we might need it.

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>